// ui/valset-standardizer-frontend/packages/src/api/generated/valset/schemas/transferObjectStatusCountViewDTO.ts
/**
 * Generated by orval v6.31.0 🍺
 * Do not edit manually.
 * Java 外部估值表标准化工程，基于 Spring Boot 和 Quartz
 * OpenAPI spec version: v1.0.0
 */

export interface TransferObjectStatusCountViewDTO {
  /** 数量。 */
  count?: number;
  /** 文件状态。 */
  status?: string;
  /** 状态名称。 */
  statusLabel?: string;
}